Limit report_jobs to only display for current user is not an admin and authenticated
